+++ src/mappings/UInt32Mapping.cpp
@@ -25,7 +25,7 @@
 namespace wolkabout
 {
 UInt32Mapping::UInt32Mapping(const std::string& reference, RegisterMapping::RegisterType registerType,
-                             const std::vector<int16_t>& addresses, RegisterMapping::OperationType operation,
+                             const std::vector<int32_t>& addresses, RegisterMapping::OperationType operation,
                              bool readRestricted, int16_t slaveAddress)
 : RegisterMapping(reference, registerType, addresses, OutputType::UINT32, operation, readRestricted, slaveAddress)
 {
